pyproject.toml
[tool.poetry]
name = "getpocket-api"
version = "0.1.0"
description = "Retrieve Pocket Posts by Tag Programmatically by using pocket's API"
authors = ["Argandov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
datetime = "^5.5"
requests = "^2.31.0"
argparse = "^1.4.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
